Shader problems on OpenGLES3

+++ Source/Resources/Shaders/FairyGUI-Image.shader
@@ -68,11 +68,11 @@ Shader "FairyGUI/Image"
 					float2 texcoord : TEXCOORD0;
 
 					#ifdef CLIPPED
-					half2 clipPos : TEXCOORD1;
+					float2 clipPos : TEXCOORD1;
 					#endif
 
 					#ifdef SOFT_CLIPPED
-					half2 clipPos : TEXCOORD1;
+					float2 clipPos : TEXCOORD1;
 					#endif
 				};
 	
@@ -88,7 +88,7 @@ Shader "FairyGUI/Image"
 
 				#ifdef SOFT_CLIPPED
 				float4 _ClipBox = float4(-2, -2, 0, 0);
-				half4 _ClipSoftness = half4(0, 0, 0, 0);
+				float4 _ClipSoftness = float4(0, 0, 0, 0);
 				#endif
 
 				#ifdef COLOR_FILTER
@@ -128,7 +128,7 @@ Shader "FairyGUI/Image"
 					#endif
 
 					#ifdef SOFT_CLIPPED
-					half2 factor = half2(0,0);
+					float2 factor = float2(0,0);
 					if(i.clipPos.x<0)
 						factor.x = (1.0-abs(i.clipPos.x)) * _ClipSoftness.x;
 					else
@@ -141,7 +141,7 @@ Shader "FairyGUI/Image"
 					#endif
 
 					#ifdef CLIPPED
-					half2 factor = abs(i.clipPos);
+					float2 factor = abs(i.clipPos);
 					col.a *= step(max(factor.x, factor.y), 1);
 					#endif
